+++ subway/cc-async.cu
@@ -32,7 +32,6 @@ int main(int argc, char** argv)
 		graph.label2[i] = false;
 	}
 
-
 	gpuErrorcheck(cudaMemcpy(graph.d_outDegree, graph.outDegree, graph.num_nodes * sizeof(unsigned int), cudaMemcpyHostToDevice));
 	gpuErrorcheck(cudaMemcpy(graph.d_value, graph.value, graph.num_nodes * sizeof(unsigned int), cudaMemcpyHostToDevice));
 	gpuErrorcheck(cudaMemcpy(graph.d_label1, graph.label1, graph.num_nodes * sizeof(bool), cudaMemcpyHostToDevice));
@@ -107,7 +106,7 @@ int main(int argc, char** argv)
 	
 	gpuErrorcheck(cudaMemcpy(graph.value, graph.d_value, graph.num_nodes*sizeof(uint), cudaMemcpyDeviceToHost));
 	
-	utilities::PrintResults(graph.value, min(30, graph.num_nodes));
+	utilities::PrintResults(graph.value, min(30, (uint)graph.num_nodes));
 			
 	if(arguments.hasOutput)
 		utilities::SaveResults(arguments.output, graph.value, graph.num_nodes);
